requirements.txt
myhdl
pytest
pyyaml
tabulate
click
git+https://github.com/insper-education/pytest-devlife
requests
python-slugify
unicode
